fix(callback api): remove schema def
